ai-agent/chatbot_backend/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
dnspython==2.8.0
fastapi==0.120.1
filelock==3.20.0
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.0.2
langchain-core==1.0.1
langchain-text-splitters==1.0.0
langgraph==1.0.1
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.1
langgraph-sdk==0.2.9
langsmith==0.4.38
MarkupSafe==3.0.3
motor==3.7.1
mpmath==1.3.0
mypy_extensions==1.1.0
networkx==3.5
numpy==2.3.4
ollama==0.6.0
orjson==3.11.4
ormsgpack==1.11.0
packaging==25.0
pdf2image==1.17.0
pillow==12.0.0
pydantic==2.12.3
pydantic_core==2.41.4
pymongo==4.15.3
pytesseract==0.3.13
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
requests-toolbelt==1.0.0
safetensors==0.6.2
setuptools==80.9.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.1
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.22.1
torch==2.9.0
tqdm==4.67.1
transformers==4.57.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
xxhash==3.6.0
zstandard==0.25.0
google-auth==2.42.0
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
google-api-python-client==2.106.0
playwright==1.42.0
pandas==2.3.3
selectolax==0.4.0
accelerate>=0.20.0
google-generativeai==0.8.5